Fixed infinite loop
