+++ client/src/components/FeedPage.js
@@ -4,7 +4,7 @@ import PostPage from "./PostPage";
 import FeedTable from "./FeedTable";
 import FeedMode from "./FeedMode.js";
 import PopUpModal from "./PopUpModal";
-import PostButton from "./PostButton";
+import FeedPost from "./FeedPost";
 
 class FeedPage extends React.Component {
   constructor(props) {
@@ -25,13 +25,20 @@ class FeedPage extends React.Component {
       commentCount: 0,
       comment: "",
       update: false,
+      table: [],
     };
   }
 
   successBtn = () => {
-    this.setMode(FeedMode.FEEDTABLE);
     this.setState({
       update: true,
+    });
+    this.setMode(FeedMode.FEEDTABLE);
+  };
+
+  toggleUpdate = () => {
+    this.setState({
+      update: false,
     });
   };
 
@@ -42,7 +49,7 @@ class FeedPage extends React.Component {
     });
   };
 
-  initiateMode = (data, newMode) => {
+  initiateCommentMode = (data, newMode) => {
     this.setState({
       mode: newMode,
       id: data.id,
@@ -78,11 +85,12 @@ class FeedPage extends React.Component {
           objs: obj,
         });
       });
-    return;
+    this.buildTable();
   }
 
   async componentDidUpdate() {
     if (this.state.update) {
+      console.log("Update true");
       const url = "/posts/" + this.state.headId;
       let res = await fetch(url)
         .then((response) => response.json())
@@ -92,8 +100,121 @@ class FeedPage extends React.Component {
             update: false,
           });
         });
-    }
+      this.buildTable();
+    } else console.log("Update false");
   }
+
+  buildTable = async () => {
+    console.log("building table");
+    const table = [];
+    for (let r = 0; r < this.state.objs.length; ++r) {
+      if (this.hasBuddy(this.state.objs[r].userData.userName)) {
+        const userData = this.state.objs[r].userData;
+        const roundData = this.state.objs[r].roundData;
+        const postData = this.state.objs[r].postData;
+
+        let date = postData.date;
+
+        let name;
+        let title;
+
+        if (userData.userName === this.props.userId) name = "You";
+        else name = userData.firstName;
+
+        if (postData.postType === "round") {
+          //For round
+          title =
+            name +
+            " logged a speedgolf round. " +
+            roundData.sgs +
+            " (" +
+            roundData.strokes +
+            " in " +
+            roundData.minutes +
+            ":" +
+            roundData.seconds +
+            ") on " +
+            date;
+        } else if (postData.postType === "post") {
+          title = name + " wrote a post on " + postData.date + ".";
+        } else alert("Error");
+
+        if (!roundData.isPrivate) {
+          await this.getProfilePic(userData.userName);
+          const pic = this.state.profilePic;
+
+          table.push(
+            <tr onClick={() => this.handleTableClick(r)} key={r}>
+              <td>
+                <FeedPost
+                  title={title}
+                  pic={pic}
+                  likeCount={postData.fistBumpCount}
+                  commentCount={postData.commentCount}
+                  comment={postData.comment}
+                />
+              </td>
+            </tr>
+          );
+        }
+      }
+    } //for
+    this.setState({
+      table: table,
+    });
+  };
+
+  hasBuddy = (poster) => {
+    if (poster === this.props.userId) return true;
+    for (let bIndex = 0; bIndex < this.props.buddies.length; bIndex++) {
+      if (this.props.buddies[bIndex] === poster) return true;
+    }
+    return false;
+  };
+
+  handleTableClick = (r) => {
+    const userData = this.state.objs[r].userData;
+    const roundData = this.state.objs[r].roundData;
+
+    const commentList = this.state.objs[r].comments;
+    const postData = this.state.objs[r].postData;
+
+    if (postData.postType === "round") {
+      const data = {
+        id: this.state.objs[r]._id,
+        firstName: userData.firstName,
+        sgs: roundData.sgs,
+        minutes: roundData.minutes,
+        seconds: roundData.seconds,
+        strokes: roundData.strokes,
+        type: postData.postType,
+        comments: commentList,
+        likes: postData.fistBumpCount,
+        commentCount: postData.commentcount,
+        objs: this.state.objs[r],
+      };
+      this.initiateCommentMode(data, FeedMode.FEEDCOMMENT);
+    }
+  };
+
+  /*****************************************************************
+   * Receives the userId from the specific Feed Post.
+   * We then fetch the user object based in the userId,
+   * Currently, the get route returns a string back, so we have to
+   * parse it first before using
+   ***************************************************************** */
+  getProfilePic = async (userId) => {
+    const url = "/users/" + userId;
+
+    let res = await fetch(url)
+      .then((response) => response.json())
+      .then((user) => {
+        let userObj = JSON.parse(user); //Turn string back into Object
+        this.setState({
+          profilePic: userObj.identityData.profilePic,
+        });
+      });
+  };
 
   render() {
     switch (this.state.mode) {
@@ -108,12 +229,14 @@ class FeedPage extends React.Component {
           >
             {this.state.objs !== null ? (
               <FeedTable
+                update={this.state.update}
                 setMode={this.setMode}
                 userId={this.props.userId}
                 buddies={this.props.buddies}
                 objs={this.state.objs}
-                initiateCommentMode={this.initiateMode}
                 profilePic={this.props.profilePic}
+                toggleUpdate={this.toggleUpdate}
+                table={this.state.table}
               />
             ) : null}
           </div>
